Additional plugin refactoring/restructuring
