+++ includes/lib/plugin/posts.php
@@ -4,17 +4,21 @@
  * @version $Id$
  * @author Jessica Green <jgreen@psy-dreamer.com>
  */
+namespace MangaPress\Plugin;
+
+use MangaPress\ContentType\Taxonomy;
+use MangaPress\ContentType\PostType;
 
 /**
  * MangaPress Posts class
  * Handles functionality for the Comic post-type
  *
  * @package MangaPress
- * @subpackage MangaPress_Posts
+ * @subpackage Posts
  * @author Jessica Green <jgreen@psy-dreamer.com>
  */
 
-class MangaPress_Posts
+class Posts
 {
     /**
      * Get image html
@@ -67,9 +71,9 @@ class MangaPress_Posts
     /**
      * Class for initializing custom post-type
      *
-     * @var MangaPress_PostType
+     * @var \MangaPress\ContentType\PostType
      */
-    private $_post_type = null;
+    private $post_type = null;
 
 
     /**
@@ -84,8 +88,8 @@ class MangaPress_Posts
      */
     public function __construct()
     {
-        $this->_register_post_type();
-        $this->_rewrite_rules();
+        $this->register_post_type();
+        $this->rewrite_rules();
 
         // Setup Manga+Press Post Options box
         add_action("wp_ajax_" . self::ACTION_GET_IMAGE_HTML, array($this, 'get_image_html_ajax'));
@@ -107,10 +111,10 @@ class MangaPress_Posts
      *
      * @return void
      */
-    private function _register_post_type()
+    private function register_post_type()
     {
         // register taxonomy
-        $taxonomy = new MangaPress\Taxonomy(array(
+        $taxonomy = new Taxonomy(array(
             'name'       => self::TAX_SERIES,
             'textdomain' => MP_DOMAIN,
             'singlename' => __('Series', MP_DOMAIN),
@@ -125,7 +129,7 @@ class MangaPress_Posts
             ),
         ));
 
-        $this->_post_type = new MangaPress\PostType(array(
+        $this->post_type = new PostType(array(
             'name'          => self::POST_TYPE,
             'textdomain'    => MP_DOMAIN,
             'pluralname'    => __('Comics', MP_DOMAIN),
@@ -154,7 +158,7 @@ class MangaPress_Posts
     /**
      * Add new rewrite rules for Comic post-type
      */
-    private function _rewrite_rules()
+    private function rewrite_rules()
     {
         $post_type = self::POST_TYPE;
         $slug      = $this->get_slug();
@@ -262,7 +266,7 @@ class MangaPress_Posts
             'comic-image',
             __('Comic Image', MP_DOMAIN),
             array($this, 'comic_meta_box_cb'),
-            $this->_post_type->get_name(),
+            $this->post_type->get_name(),
             'normal',
             'high'
         );
@@ -290,7 +294,7 @@ class MangaPress_Posts
     /**
      * Enqueue scripts for post-edit and post-add screens
      *
-     * @global WP_Post $post
+     * @global \WP_Post $post
      * @return void
      */
     public function enqueue_scripts()
@@ -329,7 +333,7 @@ class MangaPress_Posts
     /**
      * Modify header columns for Comic Post-type
      *
-     * @global WP_Post $post
+     * @global \\WP_Post $post
      * @param array $column
      * @return void
      */
@@ -462,7 +466,7 @@ class MangaPress_Posts
      * meta key. This is the same meta key used for the post featured image.
      *
      * @param int $post_id
-     * @param WP_Post $post
+     * @param \\WP_Post $post
      *
      * @return int
      */
@@ -474,7 +478,7 @@ class MangaPress_Posts
         if (!wp_verify_nonce(filter_input(INPUT_POST, '_insert_comic'), self::NONCE_INSERT_COMIC))
             return $post_id;
 
-        $flash_messages = MangaPress_Bootstrap::get_instance()->get_helper('flashmessage');
+        $flash_messages = Bootstrap::get_instance()->get_helper('flashmessage');
 
         $image_ID = (int)filter_input(INPUT_POST, '_mangapress_comic_image', FILTER_SANITIZE_NUMBER_INT);
         if ($image_ID) {
