+++ mangapress.php
@@ -37,27 +37,14 @@ if(preg_match('#' . basename(__FILE__) . '#', $_SERVER['PHP_SELF']))
 
 $plugin_folder = plugin_basename(dirname(__FILE__));
 
-if (!defined('MP_VERSION'))
-    define('MP_VERSION', '3.0.0');
-
-if (!defined('MP_FOLDER'))
-    define('MP_FOLDER', $plugin_folder);
-
 if (!defined('MP_ABSPATH'))
     define('MP_ABSPATH', plugin_dir_path(__FILE__));
 
 if (!defined('MP_URLPATH'))
     define('MP_URLPATH', plugin_dir_url(__FILE__));
 
-if (!defined('MP_LANG'))
-    define('MP_LANG', $plugin_folder . '/languages');
-
-if (!defined('MP_DOMAIN'))
-    define('MP_DOMAIN', 'mangapress');
-
-
 /**
- * Autoloader
+ * Autoload plugin's classes
  */
 spl_autoload_register(function ($class) {
 
@@ -65,8 +52,7 @@ spl_autoload_register(function ($class) {
     $prefix = 'MangaPress\\';
 
     // base directory for the namespace prefix
-    $base_dir = __DIR__ . '/includes/lib/';
-
+    $base_dir = MP_ABSPATH . 'includes/lib/';
     // does the class use the namespace prefix?
     $len = strlen($prefix);
     if (strncmp($prefix, $class, $len) !== 0) {
@@ -88,318 +74,24 @@ spl_autoload_register(function ($class) {
     }
 });
 
+if (!defined('MP_VERSION'))
+    define('MP_VERSION', '3.0.0');
+
+if (!defined('MP_FOLDER'))
+    define('MP_FOLDER', $plugin_folder);
+
+if (!defined('MP_LANG'))
+    define('MP_LANG', $plugin_folder . '/languages');
+
+if (!defined('MP_DOMAIN'))
+    define('MP_DOMAIN', 'mangapress');
+
 require_once MP_ABSPATH . 'includes/functions.php';
 require_once MP_ABSPATH . 'includes/template-functions.php';
-require_once MP_ABSPATH . 'mangapress-install.php';
-require_once MP_ABSPATH . 'mangapress-admin.php';
-require_once MP_ABSPATH . 'mangapress-options.php';
-require_once MP_ABSPATH . 'mangapress-posts.php';
 
-$install = MangaPress_Install::get_instance();
+$install = MangaPress\Plugin\Install::get_instance();
 
 register_activation_hook(__FILE__, array($install, 'do_activate'));
 register_deactivation_hook(__FILE__, array($install, 'do_deactivate'));
 
-add_action('plugins_loaded', array('MangaPress_Bootstrap', 'load_plugin'));
-
-/**
- * Plugin bootstrap class.
- *
- * @package MangaPress
- * @subpackage MangaPress_Bootstrap
- * @author Jess Green <jgreen@psy-dreamer.com>
- */
-class MangaPress_Bootstrap
-{
-
-
-    /**
-     * Options array
-     *
-     * @var array
-     */
-    protected $options;
-
-
-    /**
-     * Instance of MangaPress_Bootstrap
-     *
-     * @var MangaPress_Bootstrap
-     */
-    protected static $instance;
-
-
-    /**
-     * MangaPress Posts object
-     *
-     * @var \MangaPress_Posts
-     */
-    protected $posts_helper;
-
-
-    /**
-     * Admin page helper
-     *
-     * @var MangaPress_Admin
-     */
-    protected $admin_helper;
-
-
-    /**
-     * Flash Message helper
-     *
-     * @var MangaPress\FlashMessages
-     */
-    protected $flashmessages_helper;
-
-
-    /**
-     * Static function used to initialize Bootstrap
-     *
-     * @return void
-     */
-    public static function load_plugin()
-    {
-        self::$instance  = new self();
-    }
-
-
-    /**
-     * Get instance of MangaPress_Bootstrap
-     *
-     * @return MangaPress_Bootstrap
-     */
-    public static function get_instance()
-    {
-        if (null == self::$instance) {
-            self::$instance = new self();
-        }
-
-        return self::$instance;
-    }
-
-
-    /**
-     * PHP5 constructor method
-     */
-    protected function __construct()
-    {
-        load_plugin_textdomain(MP_DOMAIN, false, dirname( plugin_basename( __FILE__ ) ) . '/languages');
-
-        add_action('init', array($this, 'init'), 500);
-        add_action('widgets_init', array($this, 'widgets_init'));
-    }
-
-
-    /**
-     * Run init functionality
-     *
-     * @see init() hook
-     * @return void
-     */
-    public function init()
-    {
-        $this->set_options();
-        MangaPress_Options::init();
-        $this->posts_helper   = new MangaPress_Posts();
-        $this->admin_helper   = new MangaPress_Admin();
-        $this->flashmessages_helper = new MangaPress\FlashMessages(array(
-            'transient_name' => 'mangapress_messages'
-        ));
-
-
-        $this->load_current_options();
-
-        add_action('admin_enqueue_scripts', array($this, 'admin_enqueue_scripts'));
-
-        add_filter('single_template', 'mangapress_single_comic_template');
-        add_filter('template_include', 'mangapress_latestcomic_page_template');
-        add_filter('template_include', 'mangapress_comicarchive_page_template');
-
-        if (get_option('mangapress_upgrade') == 'yes') {
-            MangaPress_Install::get_instance()->do_upgrade();
-        }
-    }
-
-
-    /**
-     * Register widgets
-     */
-    public function widgets_init()
-    {
-        register_widget('MangaPress\WidgetCalendar');
-    }
-
-
-    /**
-     * Get a MangaPress helper
-     *
-     * @param string $helper_name Allowed values: admin, options, posts, flashmessages
-     * @return \MangaPress_Admin|\MangaPress_Options|\MangaPress_Posts|\MangaPress_FlashMessages|\WP_Error
-     */
-    public function get_helper($helper_name)
-    {
-        $helper = "{$helper_name}_helper";
-        if (property_exists($this, $helper)) {
-            return $this->$helper;
-        }
-
-        return new WP_Error('_mangapress_helper_access', 'No helper exists by that name');
-    }
-
-
-    /**
-     * Set MangaPress options. This method should run every time
-     * MangaPress options are updated.
-     *
-     * @uses init()
-     * @see MangaPress_Bootstrap::init()
-     *
-     * @return void
-     */
-    public function set_options()
-    {
-        $this->options = maybe_unserialize(get_option('mangapress_options'));
-    }
-
-
-    /**
-     * Get MangaPress options
-     *
-     * @return array
-     */
-    public function get_options()
-    {
-        return $this->options;
-    }
-
-
-    /**
-     * Get one option from options array
-     *
-     * @param string $section Option section
-     * @param string $option_name Option name
-     * @return boolean|mixed
-     */
-    public function get_option($section, $option_name)
-    {
-        if (!isset($this->options[$section][$option_name])) {
-            return false;
-        }
-
-        return $this->options[$section][$option_name];
-    }
-
-
-    /**
-     * Load current plugin options
-     *
-     * @return void
-     */
-    private function load_current_options()
-    {
-        $mp_options = $this->get_options();
-
-        /*
-         * Disable/Enable Default Navigation CSS
-         */
-        if ($mp_options['nav']['nav_css'] == 'default_css') {
-            add_action('wp_enqueue_scripts', array($this, 'wp_enqueue_scripts'));
-        }
-
-        add_action('wp_enqueue_scripts', array($this, 'wp_enqueue_other_scripts'));
-
-        /*
-         * Comic Page size
-         */
-        if ($mp_options['comic_page']['generate_comic_page']){
-            add_image_size(
-                'comic-page',
-                $mp_options['comic_page']['comic_page_width'],
-                $mp_options['comic_page']['comic_page_height'],
-                false
-            );
-        }
-
-        /*
-         * Comic Thumbnail size for Comics Listing screen
-         */
-        add_image_size('comic-admin-thumb', 60, 80, true);
-
-    }
-
-
-    /**
-     * Enqueue default navigation stylesheet
-     *
-     * @return void
-     */
-    public function wp_enqueue_scripts()
-    {
-        /*
-         * Navigation style
-         */
-        wp_register_style(
-            'mangapress-nav',
-            MP_URLPATH . 'assets/css/nav.css',
-            null,
-            MP_VERSION,
-            'screen'
-        );
-
-        wp_register_script(
-            'mangapress-bookmark',
-//            plugins_url( '/assets/js/bookmark.js', __FILE__ ),
-            MP_URLPATH . 'assets/js/bookmark.js',
-            array('jquery'),
-            MP_VERSION,
-            true
-        );
-
-        wp_enqueue_style('mangapress-nav');
-    }
-
-    public function wp_enqueue_other_scripts()
-    {
-        wp_register_script(
-            'mangapress-bookmark',
-            plugins_url( '/assets/js/bookmark.js', __FILE__ ),
-            array('jquery'),
-            MP_VERSION,
-            true
-        );
-
-        $bookmark_styles = apply_filters('mangapress_bookmark_styles', array());
-        $bookmark_localization = array(
-            'bookmarkCloseLabel' => __('close', MP_DOMAIN),
-            'bookmarkNoHistory' => __('No bookmark history available.', MP_DOMAIN),
-            'bookmarkTitle' => __('Title', MP_DOMAIN),
-            'bookmarkDate' => __('Date', MP_DOMAIN),
-        );
-
-        wp_localize_script(
-            'mangapress-bookmark',
-            strtoupper(MP_DOMAIN),
-            array_merge($bookmark_styles, $bookmark_localization)
-        );
-
-        wp_enqueue_script('mangapress-bookmark');
-    }
-
-
-    /**
-     * Enqueue admin-related styles
-     *
-     * @return void
-     */
-    public function admin_enqueue_scripts()
-    {
-        wp_enqueue_style(
-            'mangapress-icons',
-            plugins_url('assets/css/font.css', __FILE__),
-            null,
-            MP_VERSION,
-            'screen'
-        );
-    }
-}
+add_action('plugins_loaded', array('MangaPress\Plugin\Bootstrap', 'load_plugin'));
